Reinit firstMap when deleting a map
It seems fisrtMap property is set only the first time an AcMap component is created but when AcMap component is removed from view the service did not clear firstMap property. Let's say we have this code:

public ngOnDestroy(): void {
    this.mapsManagerService._removeMapById(this.mapsManagerService.getMap().getId());
    const viewer = this.mapsManagerService.getMap().getCesiumViewer();
    viewer.destroy();
  }

When Angular re-enter into the same view calling this.mapsManagerService.getMap() returns an unexisting map and any operation on the component (ex working on the viewer) generates an Exception.
With the fix I tried to get the first available element in the registered map or reset firstMap if no Components are available. In a correct angular flow it should work